+++ src/entity.h
@@ -28,17 +28,17 @@ typedef struct entity_t
     uint8_t     frameStart;
     uint8_t     frameEnd;
     uint8_t     frameYoffset;
-    uint16_t    gid;
+    double      gid;
     SDL_Texture *sprite;
     const char  *spriteImage;
     SDL_Thread  *thread;
-    int32_t     worldPosX;
-    int32_t     worldPosY;
+    double      worldPosX;
+    double      worldPosY;
 } Entity;
 
 void   entityFree(Entity *entity);
 Entity *entityInit(const char *name);
 int8_t entityLoadSprite(Entity *entity, SDL_Renderer *renderer, const char *filename);
-int8_t entityRender(SDL_Renderer *renderer, Entity *entity, int32_t posX, int32_t posY);
+int8_t entityRender(SDL_Renderer *renderer, Entity *entity, double posX, double posY);
 
 #endif
